requirements.txt
Pillow
tqdm
rectpack
